+++ mybatis-flex-core/src/main/java/com/mybatisflex/core/dialect/DialectFactory.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2022-2023, Mybatis-Flex (fuhai999@gmail.com).
+ *  Copyright (c) 2022-2025, Mybatis-Flex (fuhai999@gmail.com).
  *  <p>
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -17,10 +17,7 @@ package com.mybatisflex.core.dialect;
 
 
 import com.mybatisflex.core.FlexGlobalConfig;
-import com.mybatisflex.core.dialect.impl.CommonsDialectImpl;
-import com.mybatisflex.core.dialect.impl.DB2105Dialect;
-import com.mybatisflex.core.dialect.impl.DmDialect;
-import com.mybatisflex.core.dialect.impl.OracleDialect;
+import com.mybatisflex.core.dialect.impl.*;
 import com.mybatisflex.core.util.MapUtil;
 import com.mybatisflex.core.util.ObjectUtil;
 
@@ -41,7 +38,6 @@ public class DialectFactory {
      * 此 map 中，用于覆盖系统的方言实现
      */
     private static final Map<DbType, IDialect> dialectMap = new EnumMap<>(DbType.class);
-
     /**
      * 通过设置当前线程的数据库类型，以达到在代码执行时随时切换方言的功能
      */
@@ -112,6 +108,7 @@ public class DialectFactory {
             case DORIS:
                 return new CommonsDialectImpl(KeywordWrap.BACK_QUOTE, LimitOffsetProcessor.MYSQL);
             case CLICK_HOUSE:
+                return new ClickhouseDialectImpl(KeywordWrap.NONE, LimitOffsetProcessor.MYSQL);
             case GBASE_8S:
                 return new CommonsDialectImpl(KeywordWrap.NONE, LimitOffsetProcessor.MYSQL);
             case DM:
@@ -144,15 +141,17 @@ public class DialectFactory {
             case DB2_1005:
                 return new DB2105Dialect(KeywordWrap.NONE, DB2105Dialect.DB2105LimitOffsetProcessor.DB2105);
             case SQLSERVER:
-                return new CommonsDialectImpl(KeywordWrap.NONE_CASE_SENSITIVE, LimitOffsetProcessor.SQLSERVER);
+                return new SqlserverDialectImpl(KeywordWrap.SQUARE_BRACKETS, LimitOffsetProcessor.SQLSERVER);
             case SQLSERVER_2005:
-                return new CommonsDialectImpl(KeywordWrap.NONE_CASE_SENSITIVE, LimitOffsetProcessor.SQLSERVER_2005);
+                return new Sqlserver2005DialectImpl(KeywordWrap.SQUARE_BRACKETS, LimitOffsetProcessor.SQLSERVER_2005);
             case INFORMIX:
                 return new CommonsDialectImpl(KeywordWrap.NONE, LimitOffsetProcessor.INFORMIX);
             case SINODB:
                 return new CommonsDialectImpl(KeywordWrap.DOUBLE_QUOTATION, LimitOffsetProcessor.SINODB);
             case SYBASE:
                 return new CommonsDialectImpl(KeywordWrap.DOUBLE_QUOTATION, LimitOffsetProcessor.SYBASE);
+            case TRINO:
+                return new CommonsDialectImpl(KeywordWrap.NONE, LimitOffsetProcessor.SQLSERVER);
             default:
                 return new CommonsDialectImpl();
         }