+++ mybatis-flex-core/src/main/java/com/mybatisflex/core/service/IService.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2022-2023, Mybatis-Flex (fuhai999@gmail.com).
+ *  Copyright (c) 2022-2025, Mybatis-Flex (fuhai999@gmail.com).
  *  <p>
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -269,6 +269,21 @@ public interface IService<T> {
         return updateBatch(entities, DEFAULT_BATCH_SIZE);
     }
 
+
+    /**
+     * <p>根据数据主键批量更新数据
+     *
+     * @param entities    实体类对象集合
+     * @param ignoreNulls 是否忽略空字段
+     *                    {@code true} 表示忽略实体类中为 {@code null} 的字段，不更新这些字段。
+     *                    {@code false} 表示不忽略空字段，允许将对应字段更新为 {@code null}。
+     * @return boolean {@code true} 更新成功，{@code false} 更新失败。
+     * @apiNote 若 {@code ignoreNulls} 为 {@code true}，实体类中为 {@code null} 的属性不会更新到数据库。
+     */
+    default boolean updateBatch(Collection<T> entities, boolean ignoreNulls) {
+        return updateBatch(entities, DEFAULT_BATCH_SIZE, ignoreNulls);
+    }
+
     /**
      * <p>根据数据主键批量更新数据
      *
@@ -280,6 +295,23 @@ public interface IService<T> {
     default boolean updateBatch(Collection<T> entities, int batchSize) {
         Class<BaseMapper<T>> usefulClass = (Class<BaseMapper<T>>) ClassUtil.getUsefulClass(getMapper().getClass());
         return SqlUtil.toBool(Db.executeBatch(entities, batchSize, usefulClass, BaseMapper::update));
+    }
+
+
+    /**
+     * <p>根据数据主键批量更新数据
+     *
+     * @param entities    实体类对象集合
+     * @param batchSize   每批次更新数量
+     * @param ignoreNulls 是否忽略空字段
+     *                    {@code true} 表示忽略实体类中为 {@code null} 的字段，不更新这些字段。
+     *                    {@code false} 表示不忽略空字段，允许将对应字段更新为 {@code null}。
+     * @return {@code true} 更新成功，{@code false} 更新失败。
+     * @apiNote 若 {@code ignoreNulls} 为 {@code true}，实体类中为 {@code null} 的属性不会更新到数据库。
+     */
+    default boolean updateBatch(Collection<T> entities, int batchSize, boolean ignoreNulls) {
+        Class<BaseMapper<T>> usefulClass = (Class<BaseMapper<T>>) ClassUtil.getUsefulClass(getMapper().getClass());
+        return SqlUtil.toBool(Db.executeBatch(entities, batchSize, usefulClass, (mapper, entity) -> mapper.update(entity, ignoreNulls)));
     }
 
     // ===== 查询（查）操作 =====
@@ -297,7 +329,7 @@ public interface IService<T> {
     /**
      * <p>根据实体主键查询数据。
      *
-     * @param entity    实体对象，必须包含有主键
+     * @param entity 实体对象，必须包含有主键
      * @return 查询结果数据
      */
     default T getOneByEntityId(T entity) {
@@ -307,13 +339,14 @@ public interface IService<T> {
     /**
      * <p>根据实体主键查询数据。
      *
-     * @param entity    实体对象，必须包含有主键
+     * @param entity 实体对象，必须包含有主键
      * @return 查询结果数据
      * @apiNote 该方法会将查询结果封装为 {@link Optional} 类进行返回，方便链式操作。
      */
     default Optional<T> getByEntityIdOpt(T entity) {
         return Optional.ofNullable(getOneByEntityId(entity));
     }
+
     /**
      * <p>根据数据主键查询一条数据。
      *