+++ mybatis-flex-processor/src/main/java/com/mybatisflex/processor/builder/ContentBuilder.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2022-2023, Mybatis-Flex (fuhai999@gmail.com).
+ *  Copyright (c) 2022-2025, Mybatis-Flex (fuhai999@gmail.com).
  *  <p>
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -168,8 +168,7 @@ public class ContentBuilder {
      * 构建 Tables 文件常量属性。
      */
     public static void buildTablesField(StringBuilder importBuilder, StringBuilder fieldBuilder, TableInfo tableInfo,
-                                        String tableDefClassSuffix, String tableDefPropertiesNameStyle, String tableDefInstanceSuffix) {
-        String tableDefPackage = StrUtil.buildTableDefPackage(tableInfo.getEntityName());
+                                        String tableDefClassSuffix, String tableDefPropertiesNameStyle, String tableDefInstanceSuffix, String tableDefPackage) {
         String tableDefClassName = tableInfo.getEntitySimpleName().concat(tableDefClassSuffix);
         importBuilder.append("import ").append(tableDefPackage).append('.').append(tableDefClassName).append(";\n");
         String entityComment = tableInfo.getEntityComment();