+++ mybatis-flex-test/mybatis-flex-native-test/src/main/java/com/mybatisflex/test/Account.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2022-2023, Mybatis-Flex (fuhai999@gmail.com).
+ *  Copyright (c) 2022-2025, Mybatis-Flex (fuhai999@gmail.com).
  *  <p>
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -53,6 +53,9 @@ public class Account extends BaseEntity implements Serializable, AgeAware {
 
 //    @Column(isLogicDelete = true)
     private Boolean isDelete;
+
+    @Column(version = true)
+    private Integer version;
 
     private List<Article> articles;
 
@@ -125,6 +128,14 @@ public class Account extends BaseEntity implements Serializable, AgeAware {
         isDelete = delete;
     }
 
+    public Integer getVersion() {
+        return version;
+    }
+
+    public void setVersion(Integer version) {
+        this.version = version;
+    }
+
     public List<Article> getArticles() {
         return articles;
     }
@@ -152,6 +163,7 @@ public class Account extends BaseEntity implements Serializable, AgeAware {
             ", birthday=" + birthday +
             ", options=" + options +
             ", isDelete=" + isDelete +
+            ", version=" + version +
             ", articles=" + articles +
             ", title='" + title + '\'' +
             '}';